Organize files; Remove unused code; Move code into new files
Modify item v method
